+++ src/app/core/services/generator.service.ts
@@ -8,18 +8,18 @@ export class GeneratorService {
 
   generate(n: number): string {
     let result = '';
-    
+
     for (let i = 0; i < n; i++) {
       const charRandomPosition = Math.floor(Math.random() * this.characters.length);
       result += this.characters.charAt(charRandomPosition);
     }
-    
+
     return result;
   }
 
   constructor() { }
 }
 
-export function GeneratorFactory(n: number) {
+export function GeneratorFactory(n: number): (service: GeneratorService) => string {
   return (service: GeneratorService): string => service.generate(n);
 }