+++ src/app/shared/shared.module.ts
@@ -1,16 +1,19 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
+import { BorderDirective } from './directives/border.directive';
 
 @NgModule({
   declarations: [
-    HighlightDirective
+    HighlightDirective,
+    BorderDirective
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    HighlightDirective
+    HighlightDirective,
+    BorderDirective
   ]
 })
 export class SharedModule { }